Refactor 077.
